Don't check for maxage before processing new message
Checking for maxage at the very beginning of the process function causes
the fmu to always stay at its initial input for the maxage period until
new data is processed. This has the effect of always introducing an
initial delay of maxage in the processing - regardless of new data
potentially being available.

The fmu should only stay at its input if the current data is within
maxage AND there is no new data to process. Remove (for now by
commenting) the initial check call in the process function to provide
this semantics.

Signed-off-by: Henrik Ejersbo <hejersbo@ece.au.dk>