Add experimental priority queue for incoming content data
Add experimental priority queue implementation for incoming content data
queue. The pq implementation is optional and controlled using compile
time option (cmake option) USE_RBMQ_FMU_PRIORITY_QUEUE. Default is OFF
i.e. NOT using the pq implementation.

The pq implementation removes the use of a separate lookahead queue and
sorting of this. The incoming queue is pq and maintains sorted
order upon insertion. In some circumstances the existing lookahead queue
may grow very large due to large lookahead sizes and data waiting to be
processed. In testing this was seen to impact performance. The pq
implementation is a first attempt to try and remedy this situation.

Signed-off-by: Henrik Ejersbo <hejersbo@ece.au.dk>