Initial implementation of fmi2GetMaxStepsize
Initial implementation of fmi2GetMaxStepsize. Calculates the maximum
stepsize for each vref id and returns the minimum of those. Each max
stepsize is computed as the difference of the max message timestamp and
the current output timestamp (respecting lookahead size).

Currently, the max stepsize is only calculated and returned upon calling
fmi2GetMaxStepsize. An fmu output containing the value is to be added as
well, but not present in this commit.

Signed-off-by: Henrik Ejersbo <hejersbo@ece.au.dk>