When threaded use seperate connections for publish and consume
To avoid races when using seperate consumer threads use seperate rbmq connections for publish and consume.

This means when using both content data and health data threads, there
will be four connections: content pub+sub, and health pub+sub.

Reduce the rabbitmq socket consume (poll) timeout from 1sec to 100ms.
This to provide a little more responsive termination time for the
consumer threads.

Add channel id to the RabbitMqHandler::close method. To be able to close
the proper channel.

Allow the FmuContainer to process incoming data and check for valid
(wrt. maxage) data in the current output before blocking for new data.
Only go to the blocking if the maxage check for the current data fails
and there is no new incoming data.

Signed-off-by: Henrik Ejersbo <hejersbo@ece.au.dk>